job_survey_portfolio_flask/requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
decorator==5.2.1
duckdb==1.4.3
Flask==3.1.2
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.47.0
google-auth-oauthlib==1.2.4
gspread==5.12.4
gspread-dataframe==4.0.0
gspread-formatting==1.2.1
gspread-pandas==3.3.0
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.15.0
numpy==2.4.1
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pillow==12.1.0
protobuf==6.33.4
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sql_metadata==2.19.0
sqlparse==0.5.5
st-gsheets-connection==0.1.0
streamlit==1.53.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
validators==0.35.0
watchdog==6.0.0
Werkzeug==3.1.5
